# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API Documentation",
        "description": "API documentation for my Laravel project",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/company-certificates": {
            "get": {
                "tags": [
                    "Certificates"
                ],
                "summary": "Get all certificates",
                "operationId": "0ba3aa5f5064748c99407035706f3ac3",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Certificates"
                ],
                "summary": "Create a certificate",
                "operationId": "6d9bb5e953f4875a56c2d70df0f39988",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "certificatePhotoUrl",
                                    "startat",
                                    "endat"
                                ],
                                "properties": {
                                    "certificatePhotoUrl": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "startat": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "endat": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/company-certificates/showbycode/{id}": {
            "get": {
                "tags": [
                    "Certificates"
                ],
                "summary": "Get certificate by ID",
                "operationId": "8415190c4c8839e188dad3d9534f0ef5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/company-certificates/{id}": {
            "post": {
                "tags": [
                    "Certificates"
                ],
                "summary": "Update certificate",
                "operationId": "edbdd028c811f759d470bfcdd1495e75",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "startat": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "endat": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "certificatePhotoUrl": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Certificates"
                ],
                "summary": "Delete certificate",
                "operationId": "26444d0276109f702db92ec35af5f410",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
...
}